+++ src/core/Common.cpp
@@ -802,13 +802,13 @@ UnicodeString __fastcall ExpandEnvironmentVariables(const UnicodeString & Str)
 
   Buf.SetLength(Size);
   Buf.Unique();
-  intptr_t Len = ExpandEnvironmentStrings(Str.c_str(), const_cast<LPWSTR>(Buf.c_str()), Size);
+  intptr_t Len = ExpandEnvironmentStrings(Str.c_str(), const_cast<LPWSTR>(Buf.c_str()), (DWORD)Size);
 
   if (Len > Size)
   {
     Buf.SetLength(Len);
     Buf.Unique();
-    ExpandEnvironmentStrings(Str.c_str(), const_cast<LPWSTR>(Buf.c_str()), Len);
+    ExpandEnvironmentStrings(Str.c_str(), const_cast<LPWSTR>(Buf.c_str()), (DWORD)Len);
   }
 
   PackStr(Buf);
